Create a branch merging modularization and backend
