+++ ML4PS/neural_network/__init__.py
@@ -1,0 +1,2 @@
+from ML4PS.neural_network.fully_connected import *
+from ML4PS.neural_network.h2mgnode import *
