+++ ML4PS/neural_network/fully_connected.py
@@ -7,7 +7,7 @@ from functools import partial
 import numpy as np
 import pickle
 
-class FC:
+class Dense:
 
     def __init__(self, file=None, **kwargs):
