+++ README.md
@@ -1,2 +1,14 @@
 # MachineLearning4PowerSystems
-Toolkit to facilitate the application of Machine Learning techniques for Power Systems applications
+A toolkit to facilitate the application of Machine Learning techniques for Power Systems applications
+
+# Installation
+Directly from the github repository:
+```
+pip install git+https://github.com/bdonon/MachineLearning4PowerSystems
+```
+or
+```
+git clone https://github.com/bdonon/MachineLearning4PowerSystems
+cd MachineLearning4PowerSystems
+pip install .
+```
